mdb/requirements.txt
yamcs-pymdb==1.0.10